Update to v0.4.3
Added a global setting for only suspending the device, not locking it.
Also, improved device lock code, now uses the original methods, rather
than a GSEvent. A new bug has been found in the locking code; getting
the value of the proximity state fails, returning a NO, so a workaround
to force YES is being used instead. Also, getting the second proximity
value locks the main thread, so this needs to be run on a separate
thread.